requirements.txt
discord.py>=2.3.0
aiohttp>=3.9.0
python-dotenv>=1.0.0
psutil>=5.9.0
twitchio>=2.9.0
